--- app/src/main/java/com/community/jboss/visitingcard/BlankActivity.java
@@ -1,13 +1,0 @@
-package com.community.jboss.visitingcard;
-
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-
-public class BlankActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_blank);
-    }
-}
