+++ app/src/main/java/com/community/jboss/visitingcard/IntroScreens/SliderActivity.java
@@ -1,0 +1,37 @@
+package com.community.jboss.visitingcard.IntroScreens;
+
+import android.content.Intent;
+import android.support.design.widget.FloatingActionButton;
+import android.support.design.widget.Snackbar;
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+import android.view.View;
+
+import com.community.jboss.visitingcard.LoginActivity;
+import com.community.jboss.visitingcard.R;
+import com.community.jboss.visitingcard.VisitingCard.VisitingCardActivity;
+
+public class SliderActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_slider);
+
+        FloatingActionButton fab = findViewById(R.id.fab);
+        fab.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                Snackbar.make(view, "Proceed to Login", Snackbar.LENGTH_LONG)
+                        .setAction("Yes", new View.OnClickListener() {
+                            @Override
+                            public void onClick(View view) {
+                                Intent toLogin = new Intent(SliderActivity.this, LoginActivity.class);
+                                startActivity(toLogin);
+                            }
+                        }).show();
+            }
+        });
+
+    }
+}
