+++ app/src/main/java/com/community/jboss/visitingcard/VisitingCard/VisitingCardActivity.java
@@ -1,0 +1,15 @@
+package com.community.jboss.visitingcard.VisitingCard;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+
+import com.community.jboss.visitingcard.R;
+
+public class VisitingCardActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.visiting_card);
+    }
+}
