+++ app/src/main/res/layout/login_content.xml
@@ -5,19 +5,21 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context=".MainActivity"
-    tools:showIn="@layout/activity_main">
+    tools:context=".LoginActivity"
+    tools:showIn="@layout/login_main">
 
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:padding="24dp"
+        android:text="Firebase Login\nUsing Gmail"
         android:textAlignment="center"
         android:textSize="24sp"
-        android:padding="24dp"
-        android:text="This is a basic application with no functionalities, please follow to the GitHub repository to know more about the issues and start working on them... \n\nHappy Coding!\n\n\nhttps://github.com/JBossOutreach/visiting-card-android"
         app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintHorizontal_bias="0.502"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.345" />
 
 </android.support.constraint.ConstraintLayout>
